requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

# LLM Libraries
instructor>=0.5.0
litellm>=1.30.0
openai>=1.10.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# AWS
boto3>=1.34.0
